Metodo save, update y desactivate usuario

+++ Proyecto intellij/GestionDeBibliotecas/src/main/java/com/sibi/GestionDeBibliotecas/Usuario/Model/UsuarioDTO.java
@@ -28,7 +28,7 @@ public class UsuarioDTO {
     private String codigo;
     private Date codigoGeneradoEn;
 
-    public Long getUsuarioId() {
+    public Integer getUsuarioId() {
         return usuarioId;
     }
